agent-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.16
langchain-ollama==0.1.3
langchain-core==0.2.40
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.2
